Change semantics of ArrayLookup
ArrayLookup used to act as a lookup of an unordered set of values that
was not allowed to include nil. With the Hiera::Answer class we can now
allow nil as a real value. This changes the semantics to allow nil, and
to treat the lookup as more of total lookup of values, so duplicates are
no longer removed.