Add basic test coverage for part of the system
The changes made to the code have been done without supporting tests.
This starts the process of backfilling those tests and fixes a few
errors that were found in writing them.

These tests are not complete and do not deal with how the system should
behave in various error conditions (such as no access to the files being
read), but provide a starting point.